requirements.txt
Flask == 2.3.2
python-steam-api == 1.0.8
beautifulsoup4 == 4.12.2
python-decouple == 3.8